fix console spaming, add gradle build
