+++ gradle.properties
@@ -1,0 +1,1 @@
+serverPlugin = D:/Projects/Minecraft/TestServer/plugins
