--- plugin.yml
@@ -1,8 +1,0 @@
-name: AutoMessage
-author: shawshark
-main: me.shawshark.automessage.AutoMessager
-version: 1.0
-api-version: 1.15
-
-commands:
-  automessage:
